Add complete score functionality
This is a bit of a big commit that implements all of the score based
testing.

The expected properties, unexpected properties, and priorityThresh are
all scored now, instead of simply pass fail.

It does quite a bit of fiddling with the values passed around, to make
things slightly cleaner.

Things it doesn't implement:
* cleaning up setting up the test cases
* actually passing configurable weights to test cases (though the
  scoring methods can handle it)

TODO:
* improve error handling quite a bit
* add a few more test cases
* add more comments