Remove ExponentialBackoff module
The `request_urls` module uses the
[HTTP.Agent](https://nodejs.org/api/http.html#http_class_http_agent)
class to limit the maximum number of open HTTP sockets (by default with
no agent configured, no artificial limit on the number of sockets is
set).

This means that instead of attempting to space out the actual calling of
request(), all the request() calls can be made at once, and then they
will only actually be performed a few at a time. In my testing this is
able to completely eliminate the possibility of overloading our server
with too many requests at once, while allowing for much faster test
suite execution.

As an added (small) bonus, HTTP keepalives are now set so the overhead
of initializing a connection for each request is removed, making things
even faster.

Finally, when results are cached in Fastly, the test suite can execute
extremely fast: I've seen over 400 request/second handled just fine.

Oh, and it's a lot less code :)