Update exponential backoff parameters
0 is now acceptable for minimum delay since all requests are queued
through a single socket. However, in the case of retries it's desireable
to slow down more severely to allow the Elasticsearch cluster to finish
processing any timed-out requests, so the exponent is increased. Finally
the maximum timeout is doubled to 20 seconds since an overloaded
Elasticsearch server tends to thrash (our 4 core prodbuild server will
sometimes have a load of 20+)