Always use red for regression count color
We currently display the regression count in yellow if there are zero
regressions, and red if there are regressions.

However, in general regressions are colored red (in test output, for
example), known failures yellow, and passes/improvements green. Keeping
this consistency feels a little more clear, so change the regression
color to always be red.