feat: Add a rate limit feature
The fuzzy-tester library is designed to have the ability to run lots of
tests very quickly against a Pelias server. This is nice when the server
can handle it, but it often can't.

Some Pelias instances also have rate limits, and it is nice to respect
them.

This change adds a `-r` command line flag that takes an integer value
and uses it to set a per-second rate-limit on HTTP requests.