fix(output): Display improvements in quiet mode
This fixes a long-standing bug from all the way back in
https://github.com/pelias/fuzzy-tester/pull/96 where improvements are
not shown in quiet mode, but the title of test suites where there was an
improvement _is_ shown.

The idea of quiet mode is to _only_ show notable output, rather than all
output of ever test, so showing improvements fits with that goal.