feat: Support configuration of max TCP sockets
The fuzzy-tester has historically been designed to be very nice to
Pelias servers its testing.

Small Pelias instances can be overwhelmed by too many requests to the
point where they take a long time to recover (because Elasticsearch is
working through a backlog of requests).

As a result, the fuzzy-tester has always been limited to a _single_ HTTP
request in flight at a time. Depending on the rate limit settings
(`-r`), new requests might be sent immediately once a previous response
is received, but multiple HTTP requests could not occur.

Large Pelias instances, or instances that are far away and have high
round trip network delay, would therefore not go through acceptance
tests as fast as they potentially could.

This change adds a new parameter, `--max-sockets / -m` which allows
raising the number of TCP sockets used by the fuzzy tester. This in
effect allows configuring the maximum number of requests in flight at
once.

The default is still 1, so the fuzzy-tester will continue to be nice to
Pelias servers by default.

But it can now send requests faster to Pelias instances that are known
to be able to handle it.