fix(score): Use improvements when calculating pass rate
It turns out we were not using the number of `improvements` (tests
expected to fail that actually passed) when calculating the total number
of tests, and therefore the percentage pass rate of the complete test
suite.

This has the effect of under-reporting pass rate in situations where
there are improvements.

Also, the autocomplete output generator was not printing improvements,
so it was looking like tests could "disappear" from results across
different runs.