feat(auth): add Header mode for credentials
We have a custom server in front of our pelias with some application
level hacks. It uses http header auth. I would like to be able to use
pelias fuzzy tester to see the differences, especially in autocomplete
testing, between our raw backend pelias, and this hacked up proxied
pelias. This change allows for that.